Move workspaces modals to src/workspaces (WOR-1347).
