+++ src/workspaces/RequestAccessModal/RequestAccessModal.test.ts
@@ -3,8 +3,8 @@ import { axe } from 'jest-axe';
 import { h } from 'react-hyperscript-helpers';
 import { Ajax } from 'src/libs/ajax';
 import { azureControlledAccessRequestMessage, AzureWorkspace, GoogleWorkspace } from 'src/libs/workspace-utils';
-import { RequestAccessModal } from 'src/pages/workspaces/workspace/RequestAccessModal';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
+import { RequestAccessModal } from 'src/workspaces/RequestAccessModal/RequestAccessModal';
 
 jest.mock('src/libs/ajax');
 type AjaxContract = ReturnType<typeof Ajax>;
