+++ src/workspaces/ShareWorkspaceModal/ShareWorkspaceModal.test.ts
@@ -5,9 +5,9 @@ import { h } from 'react-hyperscript-helpers';
 import { Ajax } from 'src/libs/ajax';
 import { getTerraUser } from 'src/libs/state';
 import { AzureWorkspace, GoogleWorkspace } from 'src/libs/workspace-utils';
-import ShareWorkspaceModal from 'src/pages/workspaces/workspace/ShareWorkspaceModal/ShareWorkspaceModal';
 import { AccessEntry, RawWorkspaceAcl } from 'src/pages/workspaces/workspace/WorkspaceAcl';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
+import ShareWorkspaceModal from 'src/workspaces/ShareWorkspaceModal/ShareWorkspaceModal';
 
 jest.mock('src/libs/state', () => ({
   ...jest.requireActual('src/libs/state'),
